Add a timeout when waiting for RX
When setting some keys, we sometimes expect receive a command from the
host (e.g. caps lock). Some host send commands to set LED status during
set _or_ break.

For example, when it comes to turn on a LED host commands are sent just
after the key is pushed and nothing when released. The second time the key
is pushed no command are sent but when the key is released commands are
sent.

Corrent firmware does not handle this and always expect a command from
the host after sending some set commands. In this situation firmware is
stuck waiting for some commands that won't come.

This patch adds a timeout when waiting for commands from host. If the
command is not received after RX_TIMEOUT (which is expressed as timer
tick) its not waited anymore.

Signed-off-by: Franck Jullien <franck.jullien@odyssee-systemes.fr>